testApp/requirements.txt
Django==2.2.8
djangorestframework==3.9.1
django-cors-headers==3.11.0
